refactor reminder objects
